data_sorting_website/requirements.txt
Flask
pytesseract
Pillow
SpeechRecognition
